% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_confusion_plot.R
\name{get_confusion_plot}
\alias{get_confusion_plot}
\title{GET_CONFUSION_TBL}
\usage{
get_confusion_plot(actual, predicted, align = 1)
}
\arguments{
\item{actual}{an array}

\item{predicted}{an array}
}
\value{
a heatmap of confusion!
}
\description{
This function takes in two arrays and produces a heatmap of the confusion
matrix. The heatmap squares include proportions depending on the align parameter
of actual (\code{align} = 1) or predicted (\code{align} = 2)
}
\examples{
get_confusion_plot(c(1,2,3,3,3,3,1,1,1,2,2,3,3), c(3,2,1,3,2,3,1,1,2,1,2,3,3))
}
\keyword{confusion}
